restructured to reduce loops in dependencies
